requirements.txt
pandas
python-docx
streamlit
openpyxl